part1 of refactoring
