+++ CarRental/CarRental.Logic/Interfaces/ICustomerService.cs
@@ -1,6 +1,6 @@
 ﻿using CarRental.DAL.Models;
 
-namespace CarRental.Logic.Services;
+namespace CarRental.Logic.Interfaces;
 
 public interface ICustomerService
 {
@@ -9,6 +9,8 @@ public interface ICustomerService
     Customer? GetById(int id);
 
     void Create(Customer customer);
+
+    void Update(Customer customer);
 
     void Delete(int id);
 }
