+++ CarRental/CarRental.Logic/Services/CarService.cs
@@ -1,5 +1,6 @@
 ﻿using CarRental.DAL;
 using CarRental.DAL.Models;
+using CarRental.Logic.Interfaces;
 
 namespace CarRental.Logic.Services;
 
@@ -8,21 +9,20 @@ public class CarService : ICarService
     private static int _idCounter = CarRentalData.Cars.Max(c => c.Id);
     private List<Car> _cars = CarRentalData.Cars;
 
-    public List<Car> GetAll()
+    public IEnumerable<Car> GetAll()
     {
         return CarRentalData.Cars;
     }
 
-    public List<Car> CarByName(string name)
+    public IEnumerable<Car> GetByName(string name)
     {
-        List<Car> cars = new List<Car>();
+        List<Car> cars = new();
         if (string.IsNullOrEmpty(name))
         {
-            cars = CarRentalData.Cars;
+            cars = GetAll().ToList();
         }
         else
         {
-
             cars = CarRentalData.Cars.Where(c => c.Make.Contains(name, StringComparison.CurrentCultureIgnoreCase)
                 || c.CarModel.Contains(name, StringComparison.CurrentCultureIgnoreCase)
             ).ToList();
@@ -30,23 +30,23 @@ public class CarService : ICarService
         return cars;
     }
 
-    public List<Car> CarByYear(string read)
+    public List<Car> GetByYear(string read)
     {
         int year;
         bool makes = int.TryParse(read, out year);
-        List<Car> cars = new List<Car>();
+        List<Car> cars = new();
         if (read == null)
         {
-            cars = CarRentalData.Cars;
+            cars = GetAll().ToList();
         }
         else
         {
-            cars = CarRentalData.Cars.Where(c => c.Year == year || c.CarModel.ToLower().Contains(read.ToLower())).ToList();
+            cars = GetAll().Where(c => c.Year == year || c.CarModel.ToLower().Contains(read.ToLower())).ToList();
         }
         return cars;
     }
 
-    public List<Car> CarByAddons(string addon)
+    public List<Car> GetByAddons(string addon)
     {
         List<Car> cars = new List<Car>();
         if (string.IsNullOrEmpty(addon))
