+++ CarRental/CarRental.Logic/Services/CustomerService.cs
@@ -1,5 +1,6 @@
 ﻿using CarRental.DAL;
 using CarRental.DAL.Models;
+using CarRental.Logic.Interfaces;
 
 namespace CarRental.Logic.Services;
