+++ CarRental/CarRental.Logic/Services/RentalService.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using CarRental.Logic.Interfaces;
 
 namespace CarRental.Logic.Services
 {
