+++ CarRental/CarRental.Web/Program.cs
@@ -1,3 +1,4 @@
+using CarRental.Logic.Interfaces;
 using CarRental.Logic.Services;
 
 var builder = WebApplication.CreateBuilder(args);
